82.6% testing resnet pkl
