+++ docker/docker-compose.base.yml
@@ -21,6 +21,9 @@ services:
     env_file:
       - ./minio.env
     command: server /data --console-address ":9001"
+    ports:
+      - "9000:9000"
+      - "9001:9001"
     volumes:
       - minio-data:/data
